CTL-13 Compilation issue with template classes fixed
