+++ Makefile
@@ -1,15 +1,16 @@
 SOURCE=$(wildcard Source/*.cpp)
-OBJECTS=$(SOURCE:.cpp=.o) 
-
+TEST=$(wildcard Source/Tests/*.cpp)
+OBJECTS=$(SOURCE:.cpp=.o) $(TEST:.cpp=.o)  
+CXXFLAGS= -O3
 all:$(OBJECTS)
-	g++ -o $@ $(OBJECTS)
+	g++ $(CXXFLAGS) -o $@ $(OBJECTS)
 
 %.o:%.cpp
-	g++ -c $< -o $@
+	g++ $(CXXFLAGS) -c $< -o $@
 
 print::
 	@echo $(SOURCE)
 	@echo $(OBJECTS)
 
 clean:
-	rm -rf Source/*.o *.exe
+	rm -rf Source/*.o Source/Tests/*.o *.exe
